BaekJoon

+++ 06.BaekJoon/1629(곱셈)/1629.cpp
@@ -3,10 +3,10 @@ using namespace std;
 
 int a, b, c;
 
-int func(int x){
+long long int func(int x){
 	if(x == 1) return a % c;
 	
-	int k = func(x/2) % c;
+	long long int k = func(x/2) % c;
 	
 	if(x%2 == 0) return k*k%c;
 	else return k*k%c*a%c;
@@ -16,7 +16,7 @@ int main(void){
 	scanf("%d %d %d", &a, &b, &c);
 	
 	
-	printf("%d\n", func(b));
+	printf("%lld\n", func(b));
 	
 	return 0;
 }
